acts/requirements.txt
django==4.2.*
djangorestframework==3.15.*
psycopg2-binary==2.9.*
python-decouple==3.8
Pillow>=11.0.0
dj-database-url==3.0.*
joblib>=1.3
scikit-learn>=1.3
spacy>=3.7
fuzzywuzzy>=0.18
python-Levenshtein>=0.21
pandas>=2.0
requests>=2.31
whitenoise>=6.6